+++ src/tests/test_resultset.py
@@ -10,6 +10,7 @@ import pytest
 import pandas as pd
 import polars as pl
 import sqlalchemy
+from IPython.core.error import UsageError
 
 from sql.connection import DBAPIConnection, SQLAlchemyConnection
 from sql.run.resultset import ResultSet
@@ -757,3 +758,25 @@ INSERT INTO Cities VALUES ('US', 'New York City', 2020, 8772);
     }
     expected = getattr(library, "DataFrame")(expected_result)
     assert getattr(result, equal_func)(expected)
+
+
+@pytest.mark.parametrize(
+    "query, error",
+    [
+        (
+            "select age(TIMESTAMP '2020-12-25', TIMESTAMP '2020-12-26') age",
+            "Python int too large to convert to C int",
+        ),
+        (
+            "select format('Works {:,}', 12) ok, format('Will not work {:,}', "
+            "12.0) not_ok",
+            "Thousand separators are not supported for floating point numbers",
+        ),
+    ],
+    ids=["negative_age", "thousand_separators"],
+)
+def test_db_driver_error_raised(ip_empty, query, error):
+    ip_empty.run_cell("%sql duckdb://")
+    with pytest.raises(UsageError) as e:
+        ip_empty.run_cell(f"%sql {query}").result
+    assert error in str(e.value)